Move KSVisitorVoid from api/.../symbol to api/.../visitor
This is a breaking change since KSVisitorVoid will have to be reimported. However, it seems inconsistent to me that this is the only visitor in the symbol package, and that there is a dedicated visitor package.

I'm fully aware that we cannot easily merge this due to its breaking nature, but I wanted to see how many things were impacted by this refactoring and also to discuss when such a change would be possible.

Lastly, I did not update api.base to ensure CI fails.
