+++ test-utils/src/main/kotlin/com/google/devtools/ksp/processor/BaseVisitor.kt
@@ -3,7 +3,7 @@ package com.google.devtools.ksp.processor
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSFile
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.google.devtools.ksp.symbol.KSVisitorVoid
+import com.google.devtools.ksp.visitor.KSVisitorVoid
 
 open class BaseVisitor : KSVisitorVoid() {
     override fun visitClassDeclaration(classDeclaration: KSClassDeclaration, data: Unit) {
