+++ test-utils/src/main/kotlin/com/google/devtools/ksp/processor/HelloProcessor.kt
@@ -19,6 +19,7 @@ package com.google.devtools.ksp.processor
 
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.symbol.*
+import com.google.devtools.ksp.visitor.KSVisitorVoid
 
 class HelloProcessor : AbstractTestProcessor() {
     val results = mutableListOf<String>()
