+++ test-utils/src/main/kotlin/com/google/devtools/ksp/processor/PackageAnnotationProcessor.kt
@@ -4,7 +4,7 @@ import com.google.devtools.ksp.KspExperimental
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSFile
-import com.google.devtools.ksp.symbol.KSVisitorVoid
+import com.google.devtools.ksp.visitor.KSVisitorVoid
 
 class PackageAnnotationProcessor : AbstractTestProcessor() {
     val result = mutableListOf<String>()
